Add test file for mix pools with names and ids
Check the expected behaviour when mixing id pools with name pools.
Well right now the merge.
If you have a "PoolA" named pool. Add block of 128.
Then add a block of 512 to pool id 0:
The pool status will be:
[0] : ["PoolA"] 128 -> 512

Maybe a global setting for a trasher in a project with a preprocessing
var ? can be used to swith from a "merge" behaviour to an "expanded"
behaviour.